Enteties update
Variable types and keys are selected

+++ src/BookShareHub.Core/Domain/Entities/Book.cs
@@ -8,6 +8,6 @@ public class Book
 	public string Author { get; set; }
 	public string Language { get; set; }
 	public string? Description { get; set; }
-	public int Value { get; set; } // internal abstract currency
-	public int Price { get; set; }
+	public decimal Value { get; set; } // internal abstract currency
+	public decimal Price { get; set; }
 }
